Added basic Wesnoth-like macro functionality for YAML configs.
See Shaman.yaml for a proof of concept for multi-line and single value replacements.

This currently doesn't support substituting values that have spaces in them, nor does it do any sort of validation (for example that the arguments defined match the arguments provided), and could do with a bit of a code cleanup, but it is functional.