Generate transition layers dynamically
This ensures the appropriate settings are always synced between all layers and the base map. It does introduce
a bit of a stutter in the fade-in animation on game start, but we should be able to solve that once we get background
loading implemented for the scenario file as well.